+++ exonum-java-binding-common/src/main/java/com/exonum/binding/common/proofs/map/CheckedMapProof.java
@@ -16,7 +16,8 @@
 
 package com.exonum.binding.common.proofs.map;
 
-import com.exonum.binding.common.proofs.common.CheckedProof;
+import com.exonum.binding.common.collect.MapEntry;
+import com.exonum.binding.common.proofs.CheckedProof;
 import com.google.protobuf.ByteString;
 import java.util.Set;
 
@@ -43,7 +44,7 @@ public interface CheckedMapProof extends CheckedProof {
    * Get all leaf entries of this proof.
    * @throws IllegalStateException if the proof is not valid
    */
-  Set<MapEntry> getEntries();
+  Set<MapEntry<ByteString, ByteString>> getEntries();
 
   /**
    * Get all keys that were requested, but did not appear in this proof.